2.0.0 release!
- fully implemented light theme
- change app name from theoreticaldiary back to theoretical diary
- implemented custom alert label that disappears on its own
- reimplemented lock timeout feature
- refactored some styles to the  global folder
- changed trim functions to inline ones
- added reset settings button
- updated readme

+++ src/util/eventfilters.cpp
@@ -60,7 +60,7 @@ bool BusyFilter::eventFilter(QObject *, QEvent *event)
  * InactiveFilter class.
  * Thanks to https://www.qtcentre.org/threads/1464-Detecting-user-inaction
  */
-InactiveFilter::InactiveFilter(int const i, QObject *parent) : QObject(parent)
+InactiveFilter::InactiveFilter(qint64 const i, QObject *parent) : QObject(parent)
 {
     interval = i;
     timer = new QTimer(this);
